refactor: Move section folder
